requirements.txt
spotipy==2.4.4
cleverwrap==0.2.3.6
discord==0.0.2
discord.py==0.16.12
praw==6.1.1
prawcore==1.0.1
boto==2.49.0
mysql-connector-python==3.4.0